M2-6042: Rename files to match query
